Added length and isEmpty operators
The length operator determines the length of a list or string at
the top of the stack. If the top of the stack is a list or string
it pushes an integer type containing the length of the item. If
the top of the stack is not a string or list, an error is thrown.
The operator isEmpty is very similar in terms of usage
but it pushes a boolean based on if a string or list is empty
rather than a number indicating its length. The isEmpty operator
also throws an error if the type at the top of the stack isn't a
list or string.