Added string compatability to fpush and fpop operations
Both the operators fpop and fpush can now be used on strings
in the same way as lists. However, only characters are pushed
during a pop because strings are lists of characters. Just like
only a character can be pushed into a string, unless you use the
string concatenation operator.