Added Monadic Integration
The code now has monads integrated in it. This was done so IO could
be implemented in the future. It appears to generally work though
it appears to be freezing on some programs so it's still got some
weirdness going on. Once this weirdness is fixed, then IO will be
implemented.