Added IO in Print Form
It is now possible to print stuff in EcksDee using the print command!
The print operator takes a string at the top of the stack and prints
it, simple as. If it's not a string, an error is thrown. The stack
also can't be empty. Generally, it's now possible to finally make
a hello world program in EcksDee which is epic. Next, input will
be added, which will be *really* cool.