Reworked IO syntax
To print it's now called printLine because it prints the string
at the top of the stack and then prints a newline character.
To read a line of input it's now called readLine. I will eventually
add regular print and a read that reads all of standard in that
stil works after one read.