Added changeItemAt Operator
It is now possible to change an item at a given index of a list to
something else. To do this, you have your list on the stack, an item
to change an item in the list to, and an index to change it at.
With those three combined you can change an item at a given index of
a list to the desired item. If the index is out of range, nothing
happens to the list. Running this operator always consumes the new
item and index and pushes a potentially altered list on the stack.