Added pow Operator
An operator was added to allow exponentiation in EcksDee. Given two
Floats or two Doubles on a stack, the pow operator raises the base
the item at the second place to top of the stack to the exponent,
the item at the top of the stack. If the two types are indeed two
Floats or two Doubles and not mixed (due to the static typing) then
the values are consumed and the resultant exponentiation is produced.
For example 2.0 10.0 pow consumes the two floats and yields Float
1024.0.