Fixed Major Performance Issues
The monads in the interpretor were refactored to have fewer monads
and more efficient wrapping and unwrapping. As a result, the
performance loss due to adding IO is now almost completely diminished.
It's possible there's still a performance penalty due to IO existing
but it's now negligable. This is epic. Now it isn't horribly slow.