Made While Loops Require Boolean Type to Check.
It is now required fully for a while loop to have a boolean to
determine wheather it runs or not. If a boolean isn't at the top
of the stack, an error is thrown. This helps enforce the static
typing theme of EcksDee. Also it makes more sense because before
a loop would stup running if *anything* but a Boolean True was at
the top of the stack.