Made if statements require there to be a boolean at the top of stack
Just like while loops, if statements now require a Boolean
type at the top of the stack. If the boolean is true the true branch
runs of the if statement and if it's false then the false branch runs
or nothing happens if no else branch is provided. This helps enforce
static typing and makes the conditional stuff make more sense.