Made length function O(1) time for Lists and Strings
The list and string data types were reconfigured to store the length
in them and update as the program runs. This makes it so the length
function for both Lists and Strings runs in constant (O(1)) time.

The printStack function was also improved with the printed list
cutting off after 16 items to ensure it doesn't flood stdout if its
a very long list. The same goes for Strings though the string cutoff
is 256 characters since characters are smaller than list
items typically.