Added Objects to EcksDee
A new data type exists now in EcksDee. Objects are like lists but
they are indexed by strings of any form instead of an integer.

Objects in general mimic Objects in Javascript in how they work.

Objects have four operators to use on them. The operators are:
addField, removeField, getField, and mutateField.

addField -> Adds a field to an object with the specified name
and value. Throws an error if the field already exists.

removeField -> Removes a field from an object with a specified name.
Throws an error if the field doesn't exist in the object.

getField -> Fetches the value of a specified field from an object
and pushes it to the stack. Throws an error if the field doesn't
exist in the given object.

mutateField -> Changes the value of a specified field
in an object to the new value given. Throws an error if the
field doesn't exist or the new value's type doesn't match the old
type.