Added Basic File IO
Two operators were added that perform file IO on a basic level.
The operators are: readFile and writeFile.

readFile -> Reads the contents of a file into a string and pushes
the string onto a stack.

writeFile -> Writes a string on the stack to a file specified
via a string of its path on the stack. If the file doesn't exist,
the file is created, assuming permissions exist. This operator
replaces the contents of the file with the specified string so be
aware of that! It overwrites and doesn't append!

Other file IO operations may be added in the future.