Updated README to Account for Object Data Type Existing
The README was generally updated to include Objects in the
Data Types section as well as have the four object operators of
addField, getField, removeField, and mutateField, in the Operators
section.

Aside from these changes, there was some updates made to the final
stacks of the examples to be more up to date with how the stack
is printing now.

Also, the contains operator and the isEmpty operator work with the
Object data type now.