Added new IO Operators and Some Other Improvements
Firstly, new IO operators were added. The added
operators are: print, printChar, read,
and readChar.

print -> Like printLine but doesn't add a newline
to the end of the string being printed.

printChar -> prints a single Char to stdout
that's on top of the stack.

read -> Consumes all lines of stdin and pushes
the resulting String to the stack. Stops consuming
if the user types an extra newline or EOF is reached.

readChar -> Like read or readLine but consumes
a single Char from stdin and pushes it to the stack.

On top of new IO operators, Chars are now parsed
better than before which allows for whitespace
characters to actually be typed correctly in
code. For instance, you can actually type '\n'
and it will push a newline char to the stack
and will act like a newline char if used in IO
such as printChar. This was an annoying issue
for a long time so it's good it's fixed.

Lastly, a new operator was added called dropStack.
Essentially, dropStack clears the entire stack
regardless of what it had. This means the user
no longer has to type a bunch of drop statements
or use some kind of loop.