Added Some More Casting for Char Data Type
It is now possible to cast an Integer or
BigInteger to a Char type. If the Integer/BigInteger
is in the UTF-8 character range, then the casting
occurs, otherwise an error is thrown.

Error messages for casting strings to the numeric types
were also improved by actually showing the string that was
attempted to be cast.