Added Local Variables to EcksDee
Local variables are here! It's now possible to
use local variables instead of the old var
system. Local variables work the same
way as var but without needing the del keyword
since the variables are automatically cleaned
up when the scope they were made in is left.

Brief example:
true
while
	drop

	42
	loc mak foo ;

	loc get foo ;
	*
	loc mut foo ;
	loc get foo ;

	false
;
FINAL STACK:
Integer 1764
Integer 1764
Boolean False

In this example, the variable foo is automatically
cleaned up when the scope of the while loop resets
when the loop runs again or when the loop is finished.