Added tempStackChange to EcksDee
This block allows the user to run some code but
have the stack revert back to what it was when
the code block finishes running! This allows
for functions to be safe from stack-side-effects
as well as the ability to make variables and stuff
with less of a headache. This operator *should*
be good but it was made rather quickly so
bugs may exist. They'll get fixed if they're there
when I see them.