Fixed issue with function names!
There was an issue where if the user named their function something
not starting with a lower case, ghc would complain because it's
an undefined data constructor. To avoid this Haskell goofiness
a few lines were changed in the func def area. Super easy fix!
Now, the user can name functions whatever they want just like in the
interpreted version.